Product_Color_Mismatch_Detection-main/requirements.txt
# Core
pandas
pyyaml
requests
tqdm
Pillow

# CLIP / Transformers
torch
transformers

# LangChain + OpenAI
langchain
langchain-core
langchain-openai
openai
datasets